fix(auth): Post-review cleanup across auth pages and passkey routes
Clear stale CSRF token on fetch error, type-narrow response.json() as
unknown, extract useWebAuthnSupport to shared hook, rename useLoginCSRF
to useAuthCSRF, add client-side email format validation in passkey
signup, and clarify rate limit test description.
