+++ apps/web/src/components/auth/index.ts
@@ -3,5 +3,6 @@ export { AuthDivider } from './AuthDivider';
 export { OAuthButtons } from './OAuthButtons';
 export { GoogleOneTap } from './GoogleOneTap';
 export { MagicLinkForm } from './MagicLinkForm';
-export { PasskeySignupButton, useWebAuthnSupport } from './PasskeySignupButton';
+export { PasskeySignupButton } from './PasskeySignupButton';
+export { useWebAuthnSupport } from '@/hooks/useWebAuthnSupport';
 export { PasskeyLoginButton, useConditionalPasskeyUI } from './PasskeyLoginButton';
