+++ apps/web/src/hooks/useAuthCSRF.ts
@@ -3,7 +3,7 @@
 import { createCSRFHook } from './createCSRFHook';
 
 /**
- * Hook to manage login CSRF token for unauthenticated auth flows.
+ * Hook to manage auth CSRF token for unauthenticated auth flows.
  * Used for passkey authentication, magic link, and other public auth endpoints.
  */
-export const useLoginCSRF = createCSRFHook('/api/auth/login-csrf');
+export const useAuthCSRF = createCSRFHook('/api/auth/login-csrf');
