+++ packages/lib/src/security/__tests__/distributed-rate-limit.test.ts
@@ -374,7 +374,7 @@ describe('distributed-rate-limit', () => {
       expect(DISTRIBUTED_RATE_LIMITS.SIGNUP.windowMs).toBe(60 * 60 * 1000);
     });
 
-    it('PASSWORD_RESET has expected limits', () => {
+    it('PASSWORD_RESET allows 3 attempts per hour', () => {
       expect(DISTRIBUTED_RATE_LIMITS.PASSWORD_RESET.maxAttempts).toBe(3);
       expect(DISTRIBUTED_RATE_LIMITS.PASSWORD_RESET.windowMs).toBe(60 * 60 * 1000);
     });
